Add multicursor support to the move lines commands
